backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Skrine <tsaregorodtsev03@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
opencv-python = "^4.11.0.86"
numpy = "^2.3.1"
trimesh = "^4.6.12"
shapely = "^2.1.1"
pillow = "^11.2.1"
manifold3d = "3.1.1"
mapbox-earcut = "1.0.3"
gunicorn = "^23.0.0"
celery = "^5.5.3"
redis = "^6.2.0"
django-redis = "^6.0.0"
pywavefront = "^1.3.3"
imutils = "^0.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
